Prettifies and moves files
