+++ frontend/src/components/Donations/RegisterDonationsModal.jsx
@@ -1,14 +1,12 @@
-import {Button, LoadingOverlay, TextInput, Stack, Select} from "@mantine/core";
-import {DateInput, DatePicker} from '@mantine/dates';
-import {isNotEmpty, useForm} from "@mantine/form";
+import {Button, LoadingOverlay, Select, Stack, TextInput} from "@mantine/core";
+import {DateInput} from '@mantine/dates';
+import {useForm} from "@mantine/form";
 import axios from "axios";
 import {notifications} from "@mantine/notifications";
 import {modals} from "@mantine/modals";
-import {useDisclosure} from "@mantine/hooks";
 import {useState} from "react";
 
 function RegisterDonationsModal({ onDonationCreated }) {
-    const [opened, { toggle }] = useDisclosure();
     const [loading, setLoading] = useState(false);
 
     const form = useForm({
@@ -62,52 +60,4 @@ function RegisterDonationsModal({ onDonationCreated }) {
     );
 }
 
-function GenerateReport({admin, onAdminEdited}){
-    const [opened, {toggle}] = useDisclosure();
-    const [loading, setLoading] = useState(false);
-
-    const form = useForm({
-        mode: "uncontrolled",
-
-        initialValues: {
-            startdate: '',
-            enddate: ''
-        },
-    });
-
-    /*function onSubmit(values){
-        setLoading(true);
-        axios.put(`api/admins/${user}`, {
-            user: user,
-            oldPassword: values.oldPassword,
-            newPassword: values.newPassword
-        }).then(_ => {
-            notifications.show({message: 'Informação atualizada com sucesso.'});
-            modals.closeAll();
-            if (onAdminEdited) onAdminEdited(admin);
-        }).catch(err => {
-            if (err.response?.data?.validationErrors){
-                form.setErrors(err.response.data.validationErrors);
-            } else if (err.response?.status === 401){
-                form.setFieldError('oldPassword', 'Senha incorreta.')
-            } else {
-                console.error("Unhandled error when editing admin.", err);
-                notifications.show({message: "Erro ao editar senha.", color: 'red'});
-            }
-        }).finally(() => setLoading(false));
-    }*/
-
-    return (
-        <Stack component='form'  >
-            <LoadingOverlay visible={loading}/>
-            <DatePicker type = "range"  placeholder="" {...form.getInputProps('date')} />
-            <Button fullWidth type='submit' mt="md">
-                Gerar
-            </Button>
-        </Stack>
-    )
-}
-
-
-
-export {RegisterDonationsModal, GenerateReport};
+export default RegisterDonationsModal;
